+++ CHANGELOG.md
@@ -826,3 +826,4 @@ https://github.com/haskell/cabal/issues/6481 for more information.
     - ... List of Lists of `PVals2` ( temp `PacketVals` )
     - baby step: `OSB` & `CSB` (Open Sqr Bracket & Close ...)
     - refactor: `([]::[[PVals']])`, `PVal2` -> `PVal'`, & `Val2` -> `Val'`
+    - removed `OSB`, `,` does nothing because def action is append